stm32/qspi.c: Allow qspi_write_cmd_data to write cmd with 1 data byte.
The existing qspi for stm32 implementation can only send a spi command with
exactly 0 or 2 data bytes. Certain spiflash chips (e.g. AT25SF321B) have
commands that only take a single data byte, and will ignore the command if
more than that is sent. This allows sending a command with a single data
byte. Would be nicer to have a general function to send `n` bytes, but
there probably aren't many commands that require this.

Signed-off-by: Victor Rajewski <victor@allumeenergy.com.au>