drivers/memory/spiflash.c: Write 2nd byte of SR as separate command.
The existing spiflash driver writes both bytes of SR as a single command.
Some flash chips don't support multi-byte writes such as this, and have a
seperate command to write SR byte 1 and SR byte 2 (referred to in the code
as CR). This used in the spiflash initialisation to enable quad mode. The
quad enable bit is in the second SR byte. This change still issues the
multi-byte command, then issues a single byte command for just the second
SR byte. For chips that only support one of the commands, the unsupported
command should be ignored silently. For chips that support both types of
command, the SR will be written twice.

This depends on https://github.com/micropython/micropython/pull/11931 for
the stm32 port to allow single-byte spi commands.

Signed-off-by: Victor Rajewski <victor@allumeenergy.com.au>